+++ dusty/constants.py
@@ -40,6 +40,7 @@ VM_PERSIST_DIR = '/persist'
 VM_REPOS_DIR = '/dusty_repos'
 LOCAL_BACKUP_DIR = 'dusty-backup'
 
+DUSTY_NGINX_NAME = 'dustyInternalNginx'
 NGINX_CONFIG_DIR_IN_VM = os.path.join(VM_PERSIST_DIR, 'dustyNginx')
 NGINX_CONFIG_DIR_IN_CONTAINER = '/etc/nginx/conf.d'
 NGINX_MAX_FILE_SIZE = "500M"