+++ docs/changelog.md
@@ -2,12 +2,16 @@
 
 ## 0.6.0 (In Progress)
 
+* **Breaking**
+    * Dusty has upgraded to [Docker Machine](https://docs.docker.com/machine/)! This replaces boot2docker as the tool for managing Docker-enabled VMs.
+    * The setup override flag `--boot2docker_vm_memory` has been renamed to `--vm_memory`
 * **New**
     * Managed repos are now updated in parallel during `dusty up` and `dusty test`.
 * **Misc**
     * Dusty now uses the name `dustyInternalNginx` for its own nginx container, to make naming conflicts less likely.
 
 ## 0.5.0 (August 25, 2015)
+=======
 
 * **Breaking**
     * Dusty now uses NFS instead of rsync to get repository code from your host Mac to running containers.